Major code refactor
- modularize the functionality between codes.
- Now, Given an initial grid and a function file, we can produce the output grid with cleaner API.
- CSG has cleaner API.
