+++ CMakeLists.txt
@@ -22,24 +22,46 @@ target_link_libraries(adaptive_mesh_refinement
     Eigen3::Eigen
     mtet::mtet 
     convex_hull_membership
+    implicit_functions::implicit_functions
     nlohmann_json::nlohmann_json
-    smallvector::smallvector)
+    smallvector::smallvector
+    unordered_dense::unordered_dense)
 
-add_executable(isosurfacing "app/isosurfacing.cpp")
+add_executable(gridgen "app/gridgen")
 
-#if(${CMAKE_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_SOURCE_DIR})
-#  src(FeatureSummary)
-#  feature_summary(WHAT ALL)
-#endif()
-
-target_compile_features(isosurfacing PRIVATE cxx_std_20)
-target_include_directories(isosurfacing PUBLIC ${CMAKE_CURRENT_LIST_DIR}/src)
-target_link_libraries(isosurfacing PRIVATE
+target_compile_features(gridgen PRIVATE cxx_std_20)
+target_include_directories(gridgen PUBLIC ${CMAKE_CURRENT_LIST_DIR}/src)
+target_link_libraries(gridgen PRIVATE
     adaptive_mesh_refinement
     CLI11::CLI11
     mtet::mtet 
-    implicit_functions::implicit_functions
     convex_hull_membership
     unordered_dense::unordered_dense
     smallvector::smallvector
     )
+
+
+option(GRID_GEN_TESTS "Build adaptive grid gen unit tests" ON)
+if (GRID_GEN_TESTS)
+ include(CTest)
+ enable_testing()
+ include(Catch2)
+
+ file(GLOB TEST_FILES "${CMAKE_CURRENT_LIST_DIR}/tests/*.cpp")
+ add_executable(grid_gen_tests ${TEST_FILES})
+ target_link_libraries(grid_gen_tests
+    PRIVATE 
+    adaptive_mesh_refinement
+    Catch2::Catch2
+    mtet::mtet
+    unordered_dense::unordered_dense
+    smallvector::smallvector)
+ target_compile_features(grid_gen_tests PRIVATE cxx_std_20)
+ target_compile_definitions(grid_gen_tests PRIVATE CATCH_CONFIG_ENABLE_BENCHMARKING)
+target_compile_definitions(
+    grid_gen_tests
+    PRIVATE                   
+    -DTEST_FILE="${CMAKE_CURRENT_LIST_DIR}/data"
+)
+ catch_discover_tests(grid_gen_tests)
+endif()
