+++ cmake/Catch2.cmake
@@ -1,0 +1,18 @@
+include_guard()
+
+if (NOT TARGET Catch2::Catch2)
+    FetchContent_Declare(
+        Catch2
+        GIT_REPOSITORY https://github.com/catchorg/Catch2.git
+        GIT_TAG        v2.13.6
+        GIT_SHALLOW TRUE
+    )
+
+    FetchContent_GetProperties(Catch2)
+    if(NOT catch2_POPULATED)
+        FetchContent_Populate(Catch2)
+        add_subdirectory(${catch2_SOURCE_DIR} ${catch2_BINARY_DIR})
+        list(APPEND CMAKE_MODULE_PATH ${catch2_SOURCE_DIR}/contrib)
+        include(Catch)
+    endif()
+endif()
