+++ data/Figure21/fig21a.sh
@@ -2,4 +2,4 @@ cd ../../build/Release
 
 chmod +x isosurfacing
 
-./isosurfacing ../../data/Figure21/grid_1.json ../../data/Figure21/csg_examples_3.json -t 0.03 -o "CSG" --tree ../../data/Figure21/csg_examples_3_tree.json
+./gridgen ../../data/Figure21/grid_1.json ../../data/Figure21/csg_examples_3.json -t 0.03 -o "CSG" --tree ../../data/Figure21/csg_examples_3_tree.json
