+++ data/function_examples/18-sphere_copy.json
@@ -1,0 +1,164 @@
+[
+	{
+		"type":"sphere",
+		"center":[
+			0.5000863209684059,
+			0.49997303445132185,
+			0.5001223178009573
+		],
+		"radius":0.24964812381654017
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.6248596464108361,
+			0.5001240543163638,
+			0.5002167746677744
+		],
+		"radius":0.2504254725352702
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5001077038927806,
+			0.6248492139570598,
+			0.49999195641010047
+		],
+		"radius":0.25033698367120205
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5001956839887345,
+			0.5000661294299564,
+			0.6250745908302835
+		],
+		"radius":0.24992060126123422
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.39995610279854005,
+			0.39983256315600824,
+			0.3999683587195556
+		],
+		"radius":0.24973891283103014
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.40002869362449317,
+			0.3999294231871348,
+			0.6001218828548991
+		],
+		"radius":0.2503229364843048
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.4000584884858513,
+			0.6000994616355266,
+			0.40024367648110515
+		],
+		"radius":0.24986649382240708
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.4000841145503839,
+			0.6002425085364351,
+			0.5998540751687793
+		],
+		"radius":0.24960377165556955
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.600187868039394,
+			0.4001123587316094,
+			0.39996830040949233
+		],
+		"radius":0.24957953504315686
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5999368709855876,
+			0.39993098976148245,
+			0.6001439824718386
+		],
+		"radius":0.24966250424561634
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.6001398822262679,
+			0.5999748862165071,
+			0.4000299330125938
+		],
+		"radius":0.25047412588706014
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5999093951497778,
+			0.59976466379739,
+			0.6001667765621918
+		],
+		"radius":0.25008527244328066
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.3998177824747407,
+			0.5000468569409583,
+			0.49994900944619325
+		],
+		"radius":0.2503338546147806
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5999002801986768,
+			0.5002187593835816,
+			0.5000387209078968
+		],
+		"radius":0.25036951452256734
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5001708404105658,
+			0.4001889463065761,
+			0.4997507797085935
+		],
+		"radius":0.24981243968140043
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.4999222819525587,
+			0.6001240829010057,
+			0.5002255891248177
+		],
+		"radius":0.24985490871616264
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5000494638167944,
+			0.49991802983983313,
+			0.40017103064450965
+		],
+		"radius":0.25038555515262634
+	},
+	{
+		"type":"sphere",
+		"center":[
+			0.5002437005698794,
+			0.4997500680122914,
+			0.6002129867589107
+		],
+		"radius":0.2498959486596395
+	}
+]
