+++ src/csg.hpp
@@ -1,0 +1,13 @@
+//
+//  csg.hpp
+//  adaptive_mesh_refinement
+//
+//  Created by Yiwen Ju on 8/1/24.
+//
+
+#ifndef csg_hpp
+#define csg_hpp
+
+#include <stdio.h>
+
+#endif /* csg_hpp */
