+++ src/tet_quality.cpp
@@ -73,20 +73,4 @@ double tet_radius_ratio(const std::array<std::valarray<double>,4> &pts)
     return radius_ratio;
 }
 
-bool save_metrics(const std::string& filename,
-                  const std::array<std::string, 6>& tet_metric_labels,
-                  const std::valarray<double>& tet_metric)
-{
-    // assert stats_labels.size() == stats.size()
-    using json = nlohmann::json;
-    std::ofstream fout(filename.c_str(),std::ios::app);
-    //fout.open(filename.c_str(),std::ios::app);
-    json jOut;
-    for (size_t i = 0; i < tet_metric.size(); ++i) {
-        jOut[tet_metric_labels[i]] = tet_metric[i];
-    }
-    fout << jOut << std::endl;
-    fout.close();
-    return true;
-}
